--EKS-PATCH-- (aws_credentials): update ecr url validation regex
Description:
* Updates the regex for ECR URL validation to support isolated regions and includes additional testcases for these.

Upstream PR, Issue, KEP, etc. links:
* Kubernetes PR #113087 (https://github.com/kubernetes/kubernetes/pull/113087) added this change to 1.26. This patch
isn't a cherry pick, as it was created before the PR was opened.

If this patch is based on an upstream commit, how (if at all) do this patch and the upstream source differ?
* No differences.

If this patch's changes have not been added by upstream, why not?
* N/A

Other patches related to this patch:
* None

Changes made to this patch after its initial creation and reasons for these changes:
* None

Kubernetes version this patch can be dropped:
* v1.26 -- upstream includes these changes starting in this version

Signed-off-by: Jyoti Mahapatra <jyotima@amazon.com>