--EKS-PATCH-- Bump 1.23 runc version to 1.1.6
Description:
  Fixes CVE CVE-2023-27561 and CVE-2023-28642

Upstream PR, Issue, KEP, etc. links:
  Followed upstream PR instruction https://github.com/kubernetes/kubernetes/pull/117892

If this patch is based on an upstream commit, how (if at all) do this patch and the upstream source differ?
  None

If this patch's changes have not been added by upstream, why not?
  Upstream bumped the runc version for supported versions of k8 1.24-1.28

Other patches related to this patch:
  None

Changes made to this patch after its initial creation and reasons for these changes:
  None

Kubernetes version this patch can be dropped:
  None

Signed-off-by: Sajia Zafreen <szafreen@amazon.com>