--EKS-PATCH-- Parse ipv6 address before comparison (#107736)
Description:
* Fixes a bug related to ipv6 clusters.

Upstream PR, Issue, KEP, etc. links:
* Cherry pick of Kubernetes PR #107736 (https://github.com/kubernetes/kubernetes/pull/107736)
* Fixes Kubernetes Issue #107735 (https://github.com/kubernetes/kubernetes/issues/107735)

If this patch is based on an upstream commit, how (if at all) do this patch and the upstream source differ?
* Minor difference.

If this patch's changes have not been added by upstream, why not?
* N/A

Other patches related to this patch:
* None

Changes made to this patch after its initial creation and reasons for these changes:
* None

Kubernetes version this patch can be dropped:
* v1.24 -- upstream includes these changes starting in this version

Signed-off-by: Jyoti Mahapatra <jyotima@amazon.com>