--EKS-PATCH-- Make kubelet set alpha.kubernetes.io/provided-node-ip unconditionally
Cherry-pick of upstream Kubernetes PR #109794
(https://github.com/kubernetes/kubernetes/pull/109794), which has not
yet been approved and merged into upstream Kubernetes as of the latest
EKS-D 1.23 release. If the PR is approved, Kubernetes currently
intends to include it in the 1.25 release. This patch may change in
future EKS-D releases as to stay aligned with any modification to the
upstream PR on which it is based.

This fixes a node addresses flapping bug related to legacy and external
cloud controller manager migration, as described in upstream
Kubernetes issue #109793
(https://github.com/kubernetes/kubernetes/issues/109793)

From the original PR description:
  "The easier of two possible fixes for an issue where node addresses
  flap during an upgrade to an external CCM. This fix causes kubelet to
  apply the alpha.kubernetes.io/provided-node-ip annotation
  unconditionally (not only when --cloud-provider=external). This does
  not exclude a future fix involving the root cause of the issue,
  which is that kubelet and cloud-controller-manager both attempt to
  manage node addresses when --cloud-provider is not set to external."

Signed-off-by: Davanum Srinivas <davanum@gmail.com>